Instruments.py split into three files.
