+++ pymeasure/instruments/__init__.py
@@ -23,7 +23,8 @@
 #
 
 from ..errors import RangeError, RangeException
-from .instrument import Instrument, Channel
+from .channel import Channel
+from .instrument import Instrument
 from .resources import list_resources
 from .validators import discreteTruncate
