+++ pymeasure/instruments/instrument.py
@@ -25,450 +25,11 @@
 import logging
 import time
 
-from pymeasure.adapters.visa import VISAAdapter
+from .commonBase import CommonBase
+from ..adapters import VISAAdapter
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
-
-
-class DynamicProperty(property):
-    """ Class that allows managing python property behaviour in a "dynamic" fashion
-
-    The class allows passing, in addition to regular property parameters, a list of
-    runtime configurable parameters.
-    The effect is that the behaviour of fget/fset not only depends on the obj parameter, but
-    also on a set of keyword parameters with a default value.
-    These extra parameters are read from instance, if available, or left with the default value.
-    Dynamic behaviour is achieved by changing class or instance variables with special names
-    defined as `<prefix> + <property name> + <param name>`.
-
-    Code has been based on Python equivalent implementation of properties provided in the
-    python documentation `here <https://docs.python.org/3/howto/descriptor.html#properties>`_.
-
-    :param fget: class property fget parameter whose signature is expanded with a
-                 set of keyword arguments as in fget_params_list
-    :param fset: class property fget parameter whose signature is expanded with a
-                 set of keyword arguments as in fset_params_list
-    :param fdel: class property fdel parameter
-    :param doc: class property doc parameter
-    :param fget_params_list: List of parameter names that are dynamically configurable
-    :param fset_params_list: List of parameter names that are dynamically configurable
-    :param prefix: String to be prefixed to get dynamically configurable
-                   parameters.
-    """
-
-    def __init__(self, fget=None, fset=None, fdel=None, doc=None, fget_params_list=None,
-                 fset_params_list=None, prefix=""):
-        super().__init__(fget, fset, fdel, doc)
-        self.fget_params_list = () if fget_params_list is None else fget_params_list
-        self.fset_params_list = () if fset_params_list is None else fset_params_list
-        self.name = ""
-        self.prefix = prefix
-
-    def __get__(self, obj, objtype=None):
-        if obj is None:
-            # Property return itself when invoked from a class
-            return self
-        if self.fget is None:
-            raise AttributeError(f"Unreadable attribute {self.name}")
-
-        kwargs = {}
-        for attr in self.fget_params_list:
-            attr_instance_name = self.prefix + "_".join([self.name, attr])
-            if hasattr(obj, attr_instance_name):
-                kwargs[attr] = getattr(obj, attr_instance_name)
-        return self.fget(obj, **kwargs)
-
-    def __set__(self, obj, value):
-        if self.fset is None:
-            raise AttributeError(f"Can't set attribute {self.name}")
-        kwargs = {}
-        for attr in self.fset_params_list:
-            attr_instance_name = self.prefix + "_".join([self.name, attr])
-            if hasattr(obj, attr_instance_name):
-                kwargs[attr] = getattr(obj, attr_instance_name)
-        self.fset(obj, value, **kwargs)
-
-    def __set_name__(self, owner, name):
-        self.name = name
-
-
-class CommonBase:
-    """Base class for instruments and channels.
-
-    This class contains everything for pymeasure's property creator :meth:`control`
-    and its derivatives :meth:`measurement` and :meth:`setting`.
-    """
-
-    # Variable holding the list of DynamicProperty parameters that are configurable
-    # by users
-    _fget_params_list = ('get_command',
-                         'values',
-                         'map_values',
-                         'get_process',
-                         'command_process',
-                         'check_get_errors')
-
-    _fset_params_list = ('set_command',
-                         'validator',
-                         'values',
-                         'map_values',
-                         'set_process',
-                         'command_process',
-                         'check_set_errors')
-
-    # Prefix used to store reserved variables
-    __reserved_prefix = "___"
-
-    def __init__(self):
-        self._special_names = self._setup_special_names()
-
-    def _setup_special_names(self):
-        """ Return list of class/instance special names.
-
-        Compute the list of special names based on the list of
-        class attributes that are a DynamicProperty. Check also for class variables
-        with special name and copy them at instance level
-        Internal method, not intended to be accessed at user level."""
-        special_names = []
-        dynamic_params = tuple(set(self._fget_params_list + self._fset_params_list))
-        # Check whether class variables of DynamicProperty type are present
-        for obj in (self,) + self.__class__.__mro__:
-            for attr_name, attr in obj.__dict__.items():
-                if isinstance(attr, DynamicProperty):
-                    special_names += [attr_name + "_" + key for key in dynamic_params]
-        # Check if special variables are defined at class level
-        for obj in (self,) + self.__class__.__mro__:
-            for attr in obj.__dict__:
-                if attr in special_names:
-                    # Copy class special variable at instance level, prefixing reserved_prefix
-                    setattr(self, self.__reserved_prefix + attr, obj.__dict__[attr])
-        return special_names
-
-    def __setattr__(self, name, value):
-        """ Add reserved_prefix in front of special variables."""
-        if hasattr(self, '_special_names'):
-            if name in self._special_names:
-                name = self.__reserved_prefix + name
-        super().__setattr__(name, value)
-
-    def __getattribute__(self, name):
-        """ Prevent read access to variables with special names used to
-        support dynamic property behaviour."""
-        if name in ('_special_names', '__dict__'):
-            return super().__getattribute__(name)
-        if hasattr(self, '_special_names'):
-            if name in self._special_names:
-                raise AttributeError(
-                    f"{name} is a reserved variable name and it cannot be read")
-        return super().__getattribute__(name)
-
-    # Channel management
-    def add_child(self, cls, id, group="channels", prefix="ch", **kwargs):
-        """Add a child to this instance and return its index in the children list.
-
-        The newly created child may be accessed either by the index in the
-        children list or by the created attribute.
-        The fifth channel of `instrument` with id "F" has two access options:
-        :code:`instrument.channels[4] == instrument.ch_F`
-
-        :param cls: Class of the channel.
-        :param id: Child id how it is used in communication, e.g. `"A"`.
-        :param group: Name of the group of children, used for the list.
-        :param prefix: Group prefix for the attributes, e.g. `"ch"`
-            creates attribute `self.ch_A`. An underscore separates prefix from id.
-        :param \\**kwargs: Keyword arguments for the channel creator.
-        :returns: Index of this instance's channels.
-        """
-        child = cls(self, id, **kwargs)
-        group_list = getattr(self, group, [])
-        if not group_list:
-            # Add a grouplist to the parent
-            setattr(self, group, group_list)
-        group_list.append(child)
-        setattr(self, f"{prefix}_{child.id}", child)
-        return self.channels.index(child)
-
-    # Communication functions
-    def wait_for(self, query_delay=0):
-        """Wait for some time. Used by 'ask' to wait before reading.
-
-        Implement in subclass!
-
-        :param query_delay: Delay between writing and reading in seconds.
-        """
-        raise NotImplementedError("Implement in subclass!")
-
-    def ask(self, command, query_delay=0):
-        """Write a command to the instrument and return the read response.
-
-        :param command: Command string to be sent to the instrument.
-        :param query_delay: Delay between writing and reading in seconds.
-        :returns: String returned by the device without read_termination.
-        """
-        self.write(command)
-        self.wait_for(query_delay)
-        return self.read()
-
-    def values(self, command, separator=',', cast=float, preprocess_reply=None):
-        """Write a command to the instrument and return a list of formatted
-        values from the result.
-
-        :param command: SCPI command to be sent to the instrument
-        :param separator: A separator character to split the string into a list
-        :param cast: A type to cast the result
-        :param preprocess_reply: optional callable used to preprocess values
-            received from the instrument. The callable returns the processed
-            string.
-        :returns: A list of the desired type, or strings where the casting fails
-        """
-        results = str(self.ask(command)).strip()
-        if callable(preprocess_reply):
-            results = preprocess_reply(results)
-        results = results.split(separator)
-        for i, result in enumerate(results):
-            try:
-                if cast == bool:
-                    # Need to cast to float first since results are usually
-                    # strings and bool of a non-empty string is always True
-                    results[i] = bool(float(result))
-                else:
-                    results[i] = cast(result)
-            except Exception:
-                pass  # Keep as string
-        return results
-
-    def binary_values(self, command, query_delay=0, **kwargs):
-        """ Write a command to the instrument and return a numpy array of the binary data.
-
-        :param command: Command to be sent to the instrument.
-        :param query_delay: Delay between writing and reading in seconds.
-        :param kwargs: Arguments for :meth:`~pymeasure.Adapter.read_binary_values`.
-        :returns: NumPy array of values.
-        """
-        self.write(command)
-        self.wait_for(query_delay)
-        return self.read_binary_values(**kwargs)
-
-    # Property creators
-    @staticmethod
-    def control(  # noqa: C901 accept that this is a complex method
-        get_command,
-        set_command,
-        docs,
-        validator=lambda v, vs: v,
-        values=(),
-        map_values=False,
-        get_process=lambda v: v,
-        set_process=lambda v: v,
-        command_process=lambda c: c,
-        check_set_errors=False,
-        check_get_errors=False,
-        dynamic=False,
-        **kwargs
-    ):
-        """Return a property for the class based on the supplied
-        commands. This property may be set and read from the
-        instrument. See also :meth:`measurement` and :meth:`setting`.
-
-        :param get_command: A string command that asks for the value, set to `None`
-            if get is not supported (see also :meth:`setting`).
-        :param set_command: A string command that writes the value, set to `None`
-            if set is not supported (see also :meth:`measurement`).
-        :param docs: A docstring that will be included in the documentation
-        :param validator: A function that takes both a value and a group of valid values
-            and returns a valid value, while it otherwise raises an exception
-        :param values: A list, tuple, range, or dictionary of valid values, that can be used
-            as to map values if :code:`map_values` is True.
-        :param map_values: A boolean flag that determines if the values should be
-            interpreted as a map
-        :param get_process: A function that take a value and allows processing
-            before value mapping, returning the processed value
-        :param set_process: A function that takes a value and allows processing
-            before value mapping, returning the processed value
-        :param command_process: A function that takes a command and allows processing
-            before executing the command
-        :param check_set_errors: Toggles checking errors after setting
-        :param check_get_errors: Toggles checking errors after getting
-        :param dynamic: Specify whether the property parameters are meant to be changed in
-            instances or subclasses.
-
-        Example of usage of dynamic parameter is as follows:
-
-        .. code-block:: python
-
-            class GenericInstrument(Instrument):
-                center_frequency = Instrument.control(
-                    ":SENS:FREQ:CENT?;", ":SENS:FREQ:CENT %e GHz;",
-                    " A floating point property that represents the frequency ... ",
-                    validator=strict_range,
-                    # Redefine this in subclasses to reflect actual instrument value:
-                    values=(1, 20),
-                    dynamic=True  # enable changing property parameters on-the-fly
-                )
-
-            class SpecificInstrument(GenericInstrument):
-                # Identical to GenericInstrument, except for frequency range
-                # Override the "values" parameter of the "center_frequency" property
-                center_frequency_values = (1, 10) # Redefined at subclass level
-
-            instrument = SpecificInstrument()
-            instrument.center_frequency_values = (1, 6e9) # Redefined at instance level
-
-        .. warning:: Unexpected side effects when using dynamic properties
-
-        Users must pay attention when using dynamic properties, since definition of class and/or
-        instance attributes matching specific patterns could have unwanted side effect.
-        The attribute name pattern `property_param`, where `property` is the name of the dynamic
-        property (e.g. `center_frequency` in the example) and `param` is any of this method
-        parameters name except `dynamic` and `docs` (e.g. `values` in the example) has to be
-        considered reserved for dynamic property control.
-        """
-
-        def fget(self,
-                 get_command=get_command,
-                 values=values,
-                 map_values=map_values,
-                 get_process=get_process,
-                 command_process=command_process,
-                 check_get_errors=check_get_errors,
-                 ):
-            if get_command is None:
-                raise LookupError("Instrument property can not be read.")
-            vals = self.values(command_process(get_command), **kwargs)
-            if check_get_errors:
-                self.check_errors()
-            if len(vals) == 1:
-                value = get_process(vals[0])
-                if not map_values:
-                    return value
-                elif isinstance(values, (list, tuple, range)):
-                    return values[int(value)]
-                elif isinstance(values, dict):
-                    for k, v in values.items():
-                        if v == value:
-                            return k
-                    raise KeyError(f"Value {value} not found in mapped values")
-                else:
-                    raise ValueError(
-                        'Values of type `{}` are not allowed '
-                        'for Instrument.control'.format(type(values))
-                    )
-            else:
-                vals = get_process(vals)
-                return vals
-
-        def fset(self,
-                 value,
-                 set_command=set_command,
-                 validator=validator,
-                 values=values,
-                 map_values=map_values,
-                 set_process=set_process,
-                 command_process=command_process,
-                 check_set_errors=check_set_errors,
-                 ):
-
-            if set_command is None:
-                raise LookupError("Instrument property can not be set.")
-
-            value = set_process(validator(value, values))
-            if not map_values:
-                pass
-            elif isinstance(values, (list, tuple, range)):
-                value = values.index(value)
-            elif isinstance(values, dict):
-                value = values[value]
-            else:
-                raise ValueError(
-                    'Values of type `{}` are not allowed '
-                    'for Instrument.control'.format(type(values))
-                )
-            self.write(command_process(set_command) % value)
-            if check_set_errors:
-                self.check_errors()
-
-        # Add the specified document string to the getter
-        fget.__doc__ = docs
-
-        if dynamic:
-            fget.__doc__ += "(dynamic)"
-            return DynamicProperty(fget=fget, fset=fset,
-                                   fget_params_list=Instrument._fget_params_list,
-                                   fset_params_list=Instrument._fset_params_list,
-                                   prefix=Instrument.__reserved_prefix)
-        else:
-            return property(fget, fset)
-
-    @staticmethod
-    def measurement(get_command, docs, values=(), map_values=None,
-                    get_process=lambda v: v, command_process=lambda c: c,
-                    check_get_errors=False, dynamic=False, **kwargs):
-        """ Return a property for the class based on the supplied
-        commands. This is a measurement quantity that may only be
-        read from the instrument, not set.
-
-        :param get_command: A string command that asks for the value
-        :param docs: A docstring that will be included in the documentation
-        :param values: A list, tuple, range, or dictionary of valid values, that can be used
-            as to map values if :code:`map_values` is True.
-        :param map_values: A boolean flag that determines if the values should be
-            interpreted as a map
-        :param get_process: A function that take a value and allows processing
-            before value mapping, returning the processed value
-        :param command_process: A function that take a command and allows processing
-            before executing the command, for getting
-        :param check_get_errors: Toggles checking errors after getting
-        :param dynamic: Specify whether the property parameters are meant to be changed in
-            instances or subclasses. See :meth:`control` for an usage example.
-        """
-
-        return Instrument.control(get_command=get_command,
-                                  set_command=None,
-                                  docs=docs,
-                                  values=values,
-                                  map_values=map_values,
-                                  get_process=get_process,
-                                  command_process=command_process,
-                                  check_get_errors=check_get_errors,
-                                  dynamic=dynamic,
-                                  **kwargs)
-
-    @staticmethod
-    def setting(set_command, docs,
-                validator=lambda x, y: x, values=(), map_values=False,
-                set_process=lambda v: v,
-                check_set_errors=False, dynamic=False,
-                **kwargs):
-        """Return a property for the class based on the supplied
-        commands. This property may be set, but raises an exception
-        when being read from the instrument.
-
-        :param set_command: A string command that writes the value
-        :param docs: A docstring that will be included in the documentation
-        :param validator: A function that takes both a value and a group of valid values
-            and returns a valid value, while it otherwise raises an exception
-        :param values: A list, tuple, range, or dictionary of valid values, that can be used
-            as to map values if :code:`map_values` is True.
-        :param map_values: A boolean flag that determines if the values should be
-            interpreted as a map
-        :param set_process: A function that takes a value and allows processing
-            before value mapping, returning the processed value
-        :param check_set_errors: Toggles checking errors after setting
-        :param dynamic: Specify whether the property parameters are meant to be changed in
-            instances or subclasses. See :meth:`control` for an usage example.
-        """
-
-        return Instrument.control(get_command=None,
-                                  set_command=set_command,
-                                  docs=docs,
-                                  validator=validator,
-                                  values=values,
-                                  map_values=map_values,
-                                  set_process=set_process,
-                                  check_set_errors=check_set_errors,
-                                  dynamic=dynamic,
-                                  **kwargs)
 
 
 class Instrument(CommonBase):
@@ -651,75 +212,3 @@ class Instrument(CommonBase):
             return errors
         else:
             raise NotImplementedError("Non SCPI instruments require implementation in subclasses")
-
-
-class Channel(CommonBase):
-    """The base class for channel definitions.
-
-    This class supports dynamic properties like :class:`Instrument`,
-    but requires an instrument as a parent for communication.
-    The default implementation of :meth:`write` uses :code:`str.format` to
-    exchange :code:`'{ch}'` for the channel name.
-
-    :param instrument: The instrument (an instance of :class:`~pymeasure.instruments.Instrument`)
-        to which the channel belongs.
-    :param name: Name of the channel, as it is used for the communication.
-    """
-
-    def __init__(self, parent, id):
-        super().__init__()
-        self.parent = parent
-        self.id = id
-
-    # Calls to the instrument
-    def write(self, command, **kwargs):
-        """Write a string command to the instrument.
-
-        The channel id is inserted into the command and `write_termination` appended.
-
-        Subclass this method, in order to always prepend the channel id.
-
-        :param command: command string to be sent to the instrument.
-            '{ch}' is replaced by the channel id.
-        :param kwargs: Keyword arguments for the adapter.
-        """
-        self.parent.write(command.format(ch=self.id), **kwargs)
-
-    def write_bytes(self, content, **kwargs):
-        """Write the bytes `content` to the instrument."""
-        self.parent.write_bytes(content, **kwargs)
-
-    def read(self, **kwargs):
-        """Read up to (excluding) `read_termination` or the whole read buffer."""
-        return self.parent.read(**kwargs)
-
-    def read_bytes(self, count, **kwargs):
-        """Read a certain number of bytes from the instrument.
-
-        :param int count: Number of bytes to read. A value of -1 indicates to
-            read the whole read buffer.
-        :param kwargs: Keyword arguments for the adapter.
-        :returns bytes: Bytes response of the instrument (including termination).
-        """
-        return self.parent.read_bytes(count, **kwargs)
-
-    def write_binary_values(self, command, values, *args, **kwargs):
-        """Write binary values to the instrument.
-
-        :param command: Command to send.
-        :param values: The values to transmit.
-        :param \\*args, \\**kwargs: Further arguments to hand to the Adapter.
-        """
-        self.parent.write_binary_values(command.format(ch=self.id), values, *args, **kwargs)
-
-    def read_binary_values(self, **kwargs):
-        """Read binary values from the instrument."""
-        return self.parent.read_binary_values(**kwargs)
-
-    # Communication functions
-    def wait_for(self, query_delay=0):
-        """Wait for some time. Used by 'ask' to wait before reading.
-
-        :param query_delay: Delay between writing and reading in seconds.
-        """
-        self.parent.wait_for(query_delay)
